+++ pkg/configparser/standalone.go
@@ -1,10 +1,7 @@
 package configparser
 
 // Standalone is a singleton config manager that acts as a general manager for iris.
-var Standalone = NewConfigManager()
-
-// TestParser acts as a test config manager that can be used globally.
-var TestParser = NewConfigManager()
+var Standalone = NewConfigManager().(*managerImpl)
 
 func AddSource(s Source) {
 	Standalone.AddSource(s)
@@ -16,4 +13,8 @@ func Register(name, desc string, defaultValue interface{}) (*Options, error) {
 
 func Load() {
 	Standalone.Load()
+}
+
+func Reset() {
+	Standalone.Reset()
 }
