+++ pkg/configparser/standalone_test.go
@@ -2,7 +2,6 @@ package configparser
 
 import (
 	"os"
-	"regexp"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -10,23 +9,18 @@ import (
 
 const envNameRegex = "(ENV|env)+"
 
-var (
-	testSource    Source
-	testEnvSource *EnvSource
-	cfgAssert     *assert.Assertions
+// TestParser acts as a test config manager that can be used globally.
+var TestParser = NewConfigManager().(*managerImpl)
 
-	testOpts = &Options{
-		Name:        "config.options1",
-		Description: "this a mock options",
-		Manager:     TestParser,
-	}
-)
+var testOpts = &Options{
+	Name:        "config.options1",
+	Description: "this a mock options",
+	Manager:     TestParser,
+}
 
 func setupConfigTest(t *testing.T) {
 	t.Helper()
-	cfgAssert = assert.New(t)
 	addTestSource(t, &EnvSource{})
-	testSource = getTestEnvSource(t)
 }
 
 func createAndRegister(t *testing.T, name, desc string, defaultValue interface{}) error {
@@ -41,25 +35,39 @@ func addTestSource(t *testing.T, s Source) {
 	TestParser.AddSource(s)
 }
 
-func getTestEnvSource(t *testing.T) Source {
-	t.Helper()
-	envRgx := regexp.MustCompile(envNameRegex)
-	// since we want to get Env source we will check for names
-	for i := len(TestParser.sources) - 1; i >= 0; i-- {
-		source := TestParser.sources[i]
-		if val := envRgx.MatchString(source.Name()); val {
-			return source
-		}
-	}
-
-	t.Error("cannot find envsource in given manager")
-	return testEnvSource
-}
-
 func createTestEnvVars(t *testing.T, key, value string) {
 	t.Helper()
 	err := os.Setenv(key, value)
 	if err != nil {
 		t.Errorf("error creating envars %s: %s", key, err.Error())
 	}
+}
+
+func TestRegister(t *testing.T) {
+	t.Run("register an unvalid options to default config manager", func(t *testing.T) {
+		opt, err := Register("test-asdf", "this shouldn't register", nil)
+		assert.Error(t, err)
+		assert.Nil(t, opt)
+	})
+}
+
+func TestLoad(t *testing.T) {
+	t.Run("mock load", func(t *testing.T) {
+		// we didn't actually have any config loaded so len(options) = 0
+		Load()
+		assert.Equal(t, len(Standalone.Options), 0)
+	})
+}
+
+func TestAddSource(t *testing.T) {
+	t.Run("add envsources", func(t *testing.T) {
+		AddSource(&EnvSource{})
+		assert.Equal(t, len(Standalone.sources), 1)
+	})
+	t.Run("reset after source", func(t *testing.T) {
+		AddSource(&EnvSource{})
+		AddSource(&YamlSource{})
+		Reset()
+		assert.Equal(t, len(Standalone.sources), 0)
+	})
 }
